move to new feeds test and add to whats new
